Added handling of where test is hosted.

+++ yuniql-platformtests/Infrastructure/Bootstrapper.cs
@@ -14,27 +14,35 @@ namespace Yuniql.PlatformTests
         [AssemblyInitialize]
         public static void SetupInfrastrucuture(TestContext testContext)
         {
-            var containerFactory = new ContainerFactory();
-            var container = containerFactory.Create("sqlserver");
-
-            var dockerService = new DockerService();
-            var foundContainer = dockerService.FindByName(container.Name).FirstOrDefault();
-            if (null != foundContainer)
+            var testAgentHost = EnvironmentHelper.GetEnvironmentVariable(EnvironmentVariableNames.YUNIQL_TEST_HOST);
+            if (string.IsNullOrEmpty(testAgentHost) || string.IsNullOrWhiteSpace(testAgentHost) || testAgentHost.ToUpper().Equals("LOCAL"))
             {
-                dockerService.Remove(foundContainer);
+                var containerFactory = new ContainerFactory();
+                var container = containerFactory.Create("sqlserver");
+
+                var dockerService = new DockerService();
+                var foundContainer = dockerService.FindByName(container.Name).FirstOrDefault();
+                if (null != foundContainer)
+                {
+                    dockerService.Remove(foundContainer);
+                }
+
+                dockerService.Run(container);
+
+                Thread.Sleep(1000 * 10);
             }
-
-            dockerService.Run(container);
-
-            Thread.Sleep(1000 * 10);
         }
 
         [AssemblyCleanup]
         public static void TearDownInfrastructure()
         {
-            var dockerService = new DockerService();
-            var container = dockerService.FindByName("sqlserver-test-infra").First(); ;
-            dockerService.Remove(container);
+            var testAgentHost = EnvironmentHelper.GetEnvironmentVariable(EnvironmentVariableNames.YUNIQL_TEST_HOST);
+            if (string.IsNullOrEmpty(testAgentHost) || string.IsNullOrWhiteSpace(testAgentHost) || testAgentHost.ToUpper().Equals("LOCAL"))
+            {
+                var dockerService = new DockerService();
+                var container = dockerService.FindByName("sqlserver-test-infra").First(); ;
+                dockerService.Remove(container);
+            }
         }
 
         [TestMethod]
